Remove hacky --release option
The --release option was added to deal with branchless tempest issues. But
the real solution is for installers to provide deployer-input files. This
commit replaces the --release hack with usage of the correct method by
supplying surrogate deployer-input files for icehouse/juno in case no
installer-provided files are available.

New options from juno/kilo were added to default-overrides.conf.

Change-Id: Ie2b38fe7d86382a4bf0abbe5ce9a1401e106b413